dataCollection/requirements.txt
Mastodon.py
python-dotenv
hdfs